PROJECT/requirements.txt
streamlit==1.36.0
praw==7.7.1
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
nltk==3.9.1
altair==5.3.0